Wolf monitoring genetic samples metadata

Metadata of selected genetic samples of wolves collected between 2015 and 2021, in the scope of Slovenian National Wolf Monitoring

Usage

wolf_samples

Format

A data frame with 407 rows and 7 columns:

Sample

Sample unique identifier code

Date

Date of sample collection (format: YYYY-MM-DD)

AnimalRef

Identification string for particular animal

GeneticSex

Sex of animal to which the sample belong (format: M = male, F = female)

lat

latitude (N-S) of the sample (CRS: WGS84; EPSG: 4326)

lng

longitude (W-E) of the sample (CRS: WGS84; EPSG: 4326)

SType

Type of the sample. (Direct Saliva, Scat, Urine, Saliva, Tissue, Decomposing Tissue, Blood)

Source

Slovenian National Wolf Monitoring